fix timing

+++ intro.c
@@ -416,7 +416,7 @@ void entrypoint(void) {
 		MMTIME mmtime;
 		mmtime.wType = TIME_BYTES;
 		CHECK(waveOutGetPosition(hWaveOut, &mmtime, sizeof(mmtime)));
-		const float t = mmtime.u.cb / sizeof(SAMPLE_TYPE) / SAMPLE_RATE / 2;
+		const float t = (float)mmtime.u.cb / sizeof(SAMPLE_TYPE) / SAMPLE_RATE / 2;
 		const int itime = mmtime.u.cb;/// (float)(SAMPLES_PER_TICK);
 		//uniforms[0] = itime / (float)(SAMPLES_PER_TICK);
 		//const float time_ticks = (float)mmtime.u.cb / (BYTES_PER_TICK);
